i.ortho.rectify: use time_t where needed, else int
git-svn-id: https://svn.osgeo.org/grass/grass/trunk@71274 15284696-431f-4ddb-bdfa-cd5b030d7da7

+++ imagery/i.ortho.photo/i.ortho.rectify/local_proto.h
@@ -36,7 +36,7 @@ block *get_block(struct cache *, int);
 void release_cache(struct cache *);
 
 /* report.c */
-int report(long, int);
+int report(int, int);
 
 /* target.c */
 int get_target(char *);
