+++ imagery/i.ortho.photo/i.ortho.rectify/report.c
@@ -1,10 +1,10 @@
 #include <grass/glocale.h>
 #include "global.h"
 
-int report(long rectify, int ok)
+int report(int rectify, int ok)
 {
     int minutes, hours;
-    long seconds;
+    int seconds;
     long ncells;
 
     G_message("%s", ok ? _("complete") : _("failed"));
@@ -20,9 +20,9 @@ int report(long rectify, int ok)
     G_verbose_message(_("%d rows, %d cols (%ld cells) completed in"),
 			target_window.rows, target_window.cols, ncells);
     if (hours)
-	G_verbose_message(_("%d:%02d:%02ld hours"), hours, minutes, seconds % 60);
+	G_verbose_message(_("%d:%02d:%02d hours"), hours, minutes, seconds % 60);
     else
-	G_verbose_message(_("%d:%02ld minutes"), minutes, seconds % 60);
+	G_verbose_message(_("%d:%02d minutes"), minutes, seconds % 60);
     if (seconds)
 	G_verbose_message(_("%.1f cells per minute"),
 			  (60.0 * ncells) / ((double)seconds));
